Move entity lists to components/list
