+++ root/area/places.tt
@@ -4,7 +4,7 @@
     [% IF places.size %]
       <div id="largemap"></div>
       [% script_manifest('area/places-map.js', { 'data-args' => map_data_args }) %]
-      [% React.embed(c, 'components/PlaceList', { places => places }) WRAPPER 'components/with-pager.tt' %]
+      [% React.embed(c, 'components/list/PlaceList', { places => places }) WRAPPER 'components/with-pager.tt' %]
     [% ELSE %]
       [% l('This area is not currently associated with any places.') %]
     [% END %]
