+++ root/area/releases.tt
@@ -4,7 +4,7 @@
     <h2>[% l('Releases') %]</h2>
 
     [%- IF releases.size -%]
-        [% React.embed(c, 'components/ReleaseList', { releases => releases, checkboxes => 'add-to-merge'})
+        [% React.embed(c, 'components/list/ReleaseList', { releases => releases, checkboxes => 'add-to-merge'})
             WRAPPER 'components/with-pager.tt' -%]
     [%- ELSE -%]
       [% l('This area is not currently associated with any releases.') %]
