+++ root/artist/recordings.tt
@@ -9,7 +9,7 @@
     [%- IF recordings.size -%]
         <form action="[% c.uri_for_action('/recording/merge_queue') %]"
             method="post">
-            [% React.embed(c, 'components/RecordingList', { recordings => recordings, checkboxes => 'add-to-merge' })
+            [% React.embed(c, 'components/list/RecordingList', { recordings => recordings, checkboxes => 'add-to-merge' })
                 WRAPPER 'components/with-pager.tt' -%]
             [% form_submit(l('Add selected recordings for merging')) WRAPPER form_row IF c.user_exists %]
         </form>
