+++ root/artist/works.tt
@@ -4,7 +4,7 @@
     [% IF works.size %]
       <form action="[% c.uri_for_action('/work/merge_queue') %]"
             method="post">
-      [% React.embed(c, 'components/WorkList', { works => works, checkboxes => 'add-to-merge', showRatings => 1 }) WRAPPER 'components/with-pager.tt' %]
+      [% React.embed(c, 'components/list/WorkList', { works => works, checkboxes => 'add-to-merge', showRatings => 1 }) WRAPPER 'components/with-pager.tt' %]
       [% form_submit(l('Add selected works for merging')) WRAPPER form_row  IF c.user_exists %]
       </form>
     [% ELSE %]
