+++ root/components/list/InstrumentList.js
@@ -9,29 +9,25 @@
 
 import React from 'react';
 
-import {withCatalystContext} from '../context';
-import ArtistListEntry
-  from '../static/scripts/common/components/ArtistListEntry';
+import {withCatalystContext} from '../../context';
+import InstrumentListEntry
+  from '../../static/scripts/common/components/InstrumentListEntry';
 
-import SortableTableHeader from './SortableTableHeader';
+import SortableTableHeader from '../SortableTableHeader';
 
 type Props = {|
   +$c: CatalystContextT,
-  +artists: $ReadOnlyArray<ArtistT>,
   +checkboxes?: string,
+  +instruments: $ReadOnlyArray<InstrumentT>,
   +order?: string,
-  +showBeginEnd?: boolean,
-  +showRatings?: boolean,
   +sortable?: boolean,
 |};
 
-const ArtistList = ({
+const InstrumentList = ({
   $c,
-  artists,
   checkboxes,
+  instruments,
   order,
-  showBeginEnd,
-  showRatings,
   sortable,
 }: Props) => (
   <table className="tbl">
@@ -46,12 +42,12 @@ const ArtistList = ({
           {sortable
             ? (
               <SortableTableHeader
-                label={l('Artist')}
+                label={l('Instrument')}
                 name="name"
                 order={order}
               />
             )
-            : l('Artist')}
+            : l('Instrument')}
         </th>
         <th>
           {sortable
@@ -64,42 +60,20 @@ const ArtistList = ({
             )
             : l('Type')}
         </th>
-        <th>
-          {sortable
-            ? (
-              <SortableTableHeader
-                label={l('Gender')}
-                name="gender"
-                order={order}
-              />
-            )
-            : l('Gender')}
-        </th>
-        <th>{l('Area')}</th>
-        {showBeginEnd ? (
-          <>
-            <th>{l('Begin')}</th>
-            <th>{l('Begin Area')}</th>
-            <th>{l('End')}</th>
-            <th>{l('End Area')}</th>
-          </>
-        ) : null}
-        {showRatings ? <th>{l('Rating')}</th> : null}
+        <th>{l('Description')}</th>
       </tr>
     </thead>
     <tbody>
-      {artists.map((artist, index) => (
-        <ArtistListEntry
-          artist={artist}
+      {instruments.map((instrument, index) => (
+        <InstrumentListEntry
           checkboxes={checkboxes}
           index={index}
-          key={artist.id}
-          showBeginEnd={showBeginEnd}
-          showRatings={showRatings}
+          instrument={instrument}
+          key={instrument.id}
         />
       ))}
     </tbody>
   </table>
 );
 
-export default withCatalystContext(ArtistList);
+export default withCatalystContext(InstrumentList);
