+++ root/components/list/SeriesList.js
@@ -9,13 +9,13 @@
 
 import React from 'react';
 
-import {withCatalystContext} from '../context';
-import loopParity from '../utility/loopParity';
+import {withCatalystContext} from '../../context';
+import loopParity from '../../utility/loopParity';
 import DescriptiveLink
-  from '../static/scripts/common/components/DescriptiveLink';
-import linkedEntities from '../static/scripts/common/linkedEntities';
+  from '../../static/scripts/common/components/DescriptiveLink';
+import linkedEntities from '../../static/scripts/common/linkedEntities';
 
-import SortableTableHeader from './SortableTableHeader';
+import SortableTableHeader from '../SortableTableHeader';
 
 type Props = {|
   +$c: CatalystContextT,
