+++ root/edit/details/merge_series.tt
@@ -2,13 +2,13 @@
   <tr>
     <th>[% l('Merge:') %]</th>
     <td>
-      [%~ React.embed(c, 'components/SeriesList', { series => edit.display_data.old }) ~%]
+      [%~ React.embed(c, 'components/list/SeriesList', { series => edit.display_data.old }) ~%]
     </td>
   </tr>
   <tr>
     <th>[% l('Into:') %]</th>
     <td>
-      [%~ React.embed(c, 'components/SeriesList', { series => [edit.display_data.new] }) ~%]
+      [%~ React.embed(c, 'components/list/SeriesList', { series => [edit.display_data.new] }) ~%]
     </td>
   </tr>
 </table>
