+++ root/instrument/releases.tt
@@ -4,7 +4,7 @@
     [%- IF releases.size -%]
       <form action="[% c.uri_for_action('/release/merge_queue') %]"
             method="post">
-        [% React.embed(c, 'components/ReleaseList', { releases => releases, checkboxes => 'add-to-merge', showInstrumentCredits => 1, instrumentCredits => instrument_credits})
+        [% React.embed(c, 'components/list/ReleaseList', { releases => releases, checkboxes => 'add-to-merge', showInstrumentCredits => 1, instrumentCredits => instrument_credits})
             WRAPPER 'components/with-pager.tt' -%]
         [% form_submit(l('Add selected releases for merging')) WRAPPER form_row IF c.user_exists %]
       </form>
