+++ root/server/components.js
@@ -211,20 +211,20 @@ module.exports = {
   'artist/ArtistHeader': require('../artist/ArtistHeader'),
   'collection/CollectionHeader': require('../collection/CollectionHeader'),
   'components/Aliases': require('../components/Aliases'),
-  'components/AreaList': require('../components/AreaList'),
-  'components/ArtistList': require('../components/ArtistList'),
-  'components/EventList': require('../components/EventList'),
   'components/ExpirationTime': require('../components/ExpirationTime'),
-  'components/InstrumentList': require('../components/InstrumentList'),
-  'components/LabelList': require('../components/LabelList'),
-  'components/PlaceList': require('../components/PlaceList'),
-  'components/RecordingList': require('../components/RecordingList'),
   'components/Relationships': require('../components/Relationships'),
-  'components/ReleaseGroupList': require('../components/ReleaseGroupList'),
-  'components/ReleaseList': require('../components/ReleaseList'),
-  'components/SeriesList': require('../components/SeriesList'),
   'components/UserAccountTabs': require('../components/UserAccountTabs'),
-  'components/WorkList': require('../components/WorkList'),
+  'components/list/AreaList': require('../components/list/AreaList'),
+  'components/list/ArtistList': require('../components/list/ArtistList'),
+  'components/list/EventList': require('../components/list/EventList'),
+  'components/list/InstrumentList': require('../components/list/InstrumentList'),
+  'components/list/LabelList': require('../components/list/LabelList'),
+  'components/list/PlaceList': require('../components/list/PlaceList'),
+  'components/list/RecordingList': require('../components/list/RecordingList'),
+  'components/list/ReleaseGroupList': require('../components/list/ReleaseGroupList'),
+  'components/list/ReleaseList': require('../components/list/ReleaseList'),
+  'components/list/SeriesList': require('../components/list/SeriesList'),
+  'components/list/WorkList': require('../components/list/WorkList'),
   'edit/components/EditSidebar': require('../edit/components/EditSidebar'),
   'edit/components/EditSummary': require('../edit/components/EditSummary'),
   'edit/components/Vote': require('../edit/components/Vote'),
